Performance: Move lifecycle and subscription stuff to ReactiveElement
So that text nodes don't carry all this cruft
